HookGenerator: improve efficiency at creation time
Upon class creation time (NOT instantiation), we do the actual scanning
for hooks. We put these in a cache using a private member.

Upon actual instantiation, the methods are subsequently bound. Right now
a simple getattr is done to retrieve the method, but nothing stops us
from using partial to bind the methods ourselves. This will be
investigated later, and performance/code cleanliness/impact on
tracebacks and introspection will be evaluated.