doc: fix to be buildable on Sphinx 1.2.
Suffix lists are only supported in >=1.3.  To ensure we actually build
on the minimum version we say we require, make it a bare str instead.

Note that you still need alabaster installed so it's still difficult to
build on 1.2, but I might switch over to the rtd theme anyway, so then
it becomes moot.