Get rid of the horribad HookGenerator meta class.
Benchmarks by @awilfox show that the initialisation for hooks (what the meta
class was intended to save on, by doing it only once and doing the requisite
binding) took microseconds on his old PowerPC laptop. Other overheads literally
dwarf this, and we still have to do the binding step no matter what, so there
is little benefit in keeping this micro-optimisation around, especially when
it's rather clumsy, "magickal," and imposes weird rules on hooks.