channeltrack: fix various crashes
If we get a topic for a channel we don't know about, warn (as it
shouldn't happen under most circumstances), but don't crash.

This fixes a problem that appears to be caused by a channel not existing
and yet we're joined and getting a topic. This is probably a bug.